+++ sdk/identity/azure-identity/CHANGELOG.md
@@ -10,6 +10,8 @@
 
 ### Other Changes
 
+- Create separate lists of characters that are allowed within a tenant id vs scopes in `AzureCliCredential`.
+
 ## 1.6.0-beta.3 (2023-10-12)
 
 ### Bugs Fixed